fix(phan): resolve PhanTypeMismatchArgumentNullable* baseline suppressions
Closes the null-guard cluster from #60 across 12 methods: mostly `?? ''`
/`(int)` casts or corrected docblocks for genuinely nullable values, plus
@phan-suppress-next-line annotations (matching MediaWiki core's own idiom
in EditPage::internalAttemptSavePrivate()) where control flow already
guarantees non-null but Phan can't see it.

Also fixes a real reachable PHP 8 deprecation: PFFormInputParserFunction::run
now defaults a bare (no "=value") parser-function argument to '' instead of
null, since 3 downstream calls (explode(), handleFreeParameter(), the
preg_split() on $inFormName) require a string.

Verified with Phan against the CI-pinned SMW 6.0.1/MW 1.43/PHP 8.3 combo,
plus the full PHPUnit and phpcs suites.

Part of #60.

+++ includes/PF_TemplateField.php
@@ -138,7 +138,7 @@ class PFTemplateField {
 			return;
 		}
 
-		$proptitle = Title::makeTitleSafe( SMW_NS_PROPERTY, $this->mSemanticProperty );
+		$proptitle = Title::makeTitleSafe( (int)SMW_NS_PROPERTY, $this->mSemanticProperty );
 		if ( $proptitle === null ) {
 			return;
 		}
